requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
pika==1.3.2
psycopg2-binary==2.9.9
Pillow==10.1.0
gunicorn==21.2.0
django-redis==5.4.0
django-ratelimit==4.1.0
drf-spectacular==0.26.5
django-health-check==3.17.0
